Move commonly used declaration to separate header
Fixes clang-tidy not being able to find an included header otherwise
